Create node send goal request
This code defines a ROS node that listens for incoming messages containing user-reported litter reports. When a litter report is received, the node sends a goal request to a pick up litter action server, instructing the robot to pick up the litter and place it in a receptacle. The node also defines several callback functions that are called when specific events occur, such as receiving feedback from the action server or a result from the action.